requirements.txt
streamlit
faiss-cpu
sentence-transformers
nltk
PyMuPDF
openpyxl
pandas
python-dotenv
requests
